Fix the tests to actually do "something"
The import in test/iron-file-icons.html pointed to a non-existing file, so effectively the test
didn't do anything apart from showing that the browser doesn't fail with a missing import.

Instead of just fixing the import this commit also renames the file to match the name of the
file it (supposedly) tests.